UPDATE 1/29/18
Updated the GUI to look more like the format Sharon and I liked; There is a bit of dead space between the combobox and buttons that I plan to try to take out, and I plan to also pad the text entry areas with a little spacing just to make it look less squished together. I added action listeners for all of the buttons, and I had test print outputs.  I also added a function for getting the current time stamp, which is currently called by add, remove, and edit button presses in the action listener. I  I am unsure what to DO with the timestamp once it's returned because I know that's something that has to be wired with the database, but the framework is there, however the function correctly prints the time and date.