Edit No. 3, 1/29/18
Added some plumbing for the sorting feature; Added a sorting enumeration for the "sort by" type (name, expiration, and category), and added a function that returns that enumeration based on the string that is fed in from the sort combobox. I also added the framework for those function calls to be made for each enum (eg, name enum is called, sortByName function executes) based on the enumeration value returned. It is inside of the action listener for the sort button.