Edit #4
I added a few variables for the user's textfield entries, and did just a little bit of error checking. Added confirmation dialog boxes for when the users press the button. I was thinking it would be a good idea to ask the user if they are sure they want to remove an item. So I'd have to figure out how to get that going, but I can definitely do that. I am also currently looking into how to format the date textfield to a specific format, however for now I've just assigned it to an integer.