Edit 7
Updated to take out the table that I created (in anticipation for Sumit to upload his) and that went back a little bit to Version 5 on this site; Contains the additional notes text area and the categories combobox. Removed all sort functionality since it will be done in the JTable itself.